serve images from the repository of a wiki
In order to be able to serve static files from the repository of a wiki, we had to change how we develop smeagol.
Before this change, the ui server has proxied every api request to the spring boot server:

Browser --> Webpack DevServer (Port 3000) --> Spring Boot (Port 8080)

After this change, the spring boot has to server proxy the ui requests to the ui server:

Browser --> Spring Boot (Port 8080) --> Webpack DevServer (Port 3000)

Durring the development we use now the root context path ("/" instead of "/smaegol") for ui and rest, in production we will use "/smeagol" for both.
We use the "/" path, because it seems to be hard to use a context path with (webpack-dev-server or create-react-app).
The dispatching of the requests to the right server and uri is done by the UiFilter, which delegates to the ProductionDispatcher or DevelopmentDispatcher.
This depends on the configured stage.

For the development mode we had to copy the webpackHotDevClient from react-dev-utils, to change the port of sockjs and bypass the spring boot proxy.
We have done the bypass to avoid the implementation of a websocket proxy in the spring boot application.