#83 review: renamed file, switched `var` to `const`, added early exit
- rename `util.js` to `pathUtil.js`
util is too broad for a file name as files like this tend to become grab
bags of code, encouraging to not think about business aspects.
(Sonarqube would even mention to name one-method-files after the
function)

- switched to const/added typing
`pathUtil.js` is now a typed Flow file. Even so, `var` has become a bit
of deprecated as var brings in a larger scope compared to let and const.

- early exit
The in-place variable writing was replaced with an early return
statement. There was no need to replace a given parameter variable.