#83 review: makes react-router dependency explicit
The package-lock.json already locks this version but to be safe the
package.json does now include the corresponding version as well.

Coincidentally, the package-lock.json was out of date and is going to
be updated as well.

Note:
It was it was discovered during this review, that previously the smeagol
version was mistakenly released under a wrong version. Under usual
circumstances the current smeagol version would be v0.5.3. The future
release is going to reflect this.