#173 Add Breadcrumb and button bar to page history
This commit adds a bread crumb to an existing wiki page history by reusing
existing wiki data.

This commit adds also a bar which should provide for the buttons/dropdowns:
- New page button
- Settings button
- Branch switch dropdown

The added behaviour works fine for branches where the same file exists. It will
simply switch the view to the existing file on the other branch. The behaviour
works also on files in directories.

If a file exists in one branch but not in another, the switch will be redirected
to a page creation view.