#173 Left-align interaction buttons/branch dropdown
This commit removes the CSS property justifyContent to properly align all
buttons to the left side. Otherwise on pages with only two buttons (f. i.
the directory view under `${branchname}/pages/docs/`) the second button will
spread to the right side instead of building a group with the other button.

This property has nothing to do with the right-alignment of the branch
dropdown, though. This is because the branch dropdown resides inside its own
wrapper which contains styles for a proper right alignment. The wrapper just
grabs all left space between itself and the buttons, so that there is no
space left to spread around.