#173 Middle-align branch dropdown label
The button bar seems to be somewhat bottom-aligned with the branch dropdown
while the branch dropdown's label is top aligned. This commit middle-aligns
the label in order to reduce the difference between the buttons and the
dropdown. Bottom-aligning the label would make it look weird though, that's
why middle-align. The 6px top padding takes care of the manual horizontal
alignment as previously introduced by @jbartscher